requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
kombu==5.6.2
mysqlclient==2.2.8
packaging==26.0
prompt_toolkit==3.0.52
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
